// apps/web/src/entities/generated/entity-details.gen.ts
export const detailEntities = [
  "product",
  "recipe",
  "ingredient",
  "location",
  "inventory",
  "meal",
  "ledgerParty",
  "ledgerTransfer",
  "project",
  "task",
  "vendor",
  "purchase",
  "financialAccount",
  "financialTransaction",
  "wish",
  "expense",
] as const;
export type DetailEntity = (typeof detailEntities)[number];